+++ src/app/core/api/app-player.api.ts
@@ -23,6 +23,7 @@ export class AppPlayerApi {
       .take(1)
       .subscribe((media: GoogleApiYouTubeVideoResource) => this.playVideo(media));
       // .do((media: GoogleApiYouTubeVideoResource) => this.playVideo(media));
+      debugger;
     this.queuePlaylist(playlist);
     // this.store.dispatch(this.nowPlaylistActions.selectVideo(media));
     // this.store.dispatch(new PlaylistStore.PlayPlaylistAction(playlist.id));