+++ src/app/core/store/now-playlist/now-playlist.reducer.ts
@@ -6,6 +6,7 @@ export interface NowPlaylistInterface {
   selectedId: string;
   filter: string;
   repeat: boolean;
+  playPlaylist: boolean;
 }
 
 const initialState: NowPlaylistInterface = {
@@ -13,31 +14,32 @@ const initialState: NowPlaylistInterface = {
   selectedId: '',
   filter: '',
   repeat: false,
+  playPlaylist: false
 };
 
 export function nowPlaylist(state: NowPlaylistInterface = initialState, action: Action): NowPlaylistInterface {
   switch (action.type) {
     case NowPlaylistActions.SELECT:
-      return Object.assign({}, state, { selectedId: action.payload.id });
+      return { ...state, selectedId: action.payload.id };
 
     case NowPlaylistActions.QUEUE:
-      return Object.assign({}, state, { videos: addMedia(state.videos, action.payload) });
+      return { ...state, videos: addMedia(state.videos, action.payload) };
 
     case NowPlaylistActions.QUEUE_VIDEOS:
-      return Object.assign({}, state, { videos: addMedias(state.videos, action.payload) });
+      return { ...state, videos: addMedias(state.videos, action.payload) };
 
     case NowPlaylistActions.REMOVE:
-      return Object.assign({}, state, { videos: removeMedia(state.videos, action.payload) });
+      return { ...state, videos: removeMedia(state.videos, action.payload) };
 
     // updates index by media
     case NowPlaylistActions.UPDATE_INDEX:
-      return Object.assign({}, state, { selectedId: action.payload });
+      return { ...state, selectedId: action.payload };
 
     case NowPlaylistActions.FILTER_CHANGE:
-      return Object.assign({}, state, { filter: action.payload });
+      return { ...state, filter: action.payload };
 
     case NowPlaylistActions.REMOVE_ALL:
-      return Object.assign({}, state, { videos: [], filter: '', selectedId: '' });
+      return { ...state, videos: [], filter: '', selectedId: '' };
 
     case NowPlaylistActions.SELECT_NEXT: {
       return { ...state,
@@ -46,9 +48,9 @@ export function nowPlaylist(state: NowPlaylistInterface = initialState, action: 
     }
 
     case NowPlaylistActions.SELECT_PREVIOUS:
-      return Object.assign({}, state, {
+      return { ...state,
         selectedId: selectPreviousIndex(state.videos, state.selectedId, state.filter)
-      });
+      };
 
     case NowPlaylistActions.MEDIA_ENDED:
       return selectNextOrPreviousTrack(state, state.filter);