+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v 3.5 (2017/05/06) 
+* [NEW] - added search for playlists
+* [NEW] - queue and play whole playlist 
+* [NEW] - player control bar reflects the current media photo 
+* [NEW] - color palette updated to material 
+
 ## v 3.4.1 (2017/05/06) 
 * [UPDATE] - updated album tracks parsing algorithm - now includes more rules for parsing more formats of tracks embeded in media description  
 * [UPDATE] - restyled tracks.